Create ESO208_Ass2(Q2).py
After midsem assignment, batch 2(C), Q2.  
Write a computer program for estimation of the eigenvalues.
INPUT: First it should ask for the following: size n, matrix A and approximate relative error tolerance (in %).  Then it should give a choice to the user whether (s)he wants only the largest/ smallest eigenvalue (L) or all eigenvalues (A):
•	If the response is (L), program should compute the largest eigenvalue using Power method, and the smallest eigenvalue using an Inverse Power method and the corresponding eigenvector. (In case of the singular matrix while using the inverse power method, terminate the program with a note that absolute minimum eigenvalue is zero).
•	If the response is (A), program should use QR decomposition with Gram-Schmidt algorithm (G) and then it should compute all eigenvalues.